+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## `master`
 
 - Fix sensor titles
+- Refactor Tado° service #73
+- Add `--service-id` CLI option to run only specified services
 
 ## `0.54.0`